% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcases.R
\name{bc_collect_xml}
\alias{bc_collect_xml}
\title{Fetch XML for BatchCases rows}
\usage{
bc_collect_xml(
  rows,
  parallel = TRUE,
  workers = max(1, parallel::detectCores() - 1),
  page_sleep = 0
)
}
\description{
Fetch XML for BatchCases rows
}
