% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcases.R
\name{bc_fetch}
\alias{bc_fetch}
\title{Plain BatchCases fetch following @odata.nextLink}
\usage{
bc_fetch(
  q,
  page_sleep = 0,
  sleep_after = 0,
  endpoint = "/BatchCases",
  retry = 3,
  retry_wait = 1,
  verbose = TRUE
)
}
\description{
Plain BatchCases fetch following @odata.nextLink
}
