% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcases.R
\name{bc_fetch_resumable_auto}
\alias{bc_fetch_resumable_auto}
\title{Resumable / checkpointed BatchCases fetch (rows only)
Saves to BatchCases/\if{html}{\out{<search>}}/}
\usage{
bc_fetch_resumable_auto(
  q,
  page_sleep = 0,
  sleep_after = 0,
  endpoint = "/BatchCases",
  retry = 3,
  retry_wait = 1,
  verbose = TRUE,
  base_dir = "BatchCases"
)
}
\description{
Resumable / checkpointed BatchCases fetch (rows only)
Saves to BatchCases/\if{html}{\out{<search>}}/
}
