% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcases.R
\name{bc_run}
\alias{bc_run}
\title{High-level runner like your original: fetch, (optional) xml, (optional) parse}
\usage{
bc_run(
  q,
  page_sleep = 0,
  sleep_after = 0,
  endpoint = "/BatchCases",
  retry = 3,
  retry_wait = 1,
  verbose = TRUE,
  fetch_xml = TRUE,
  parse_xml = TRUE
)
}
\description{
High-level runner like your original: fetch, (optional) xml, (optional) parse
}
