% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcases.R
\name{bc_run_resumable_auto}
\alias{bc_run_resumable_auto}
\title{Resumable BatchCases: fetch → XML → JSON, all saved under BatchCases/\if{html}{\out{<search>}}/}
\usage{
bc_run_resumable_auto(
  q,
  page_sleep = 0,
  sleep_after = 0,
  endpoint = "/BatchCases",
  retry = 3,
  retry_wait = 1,
  verbose = TRUE,
  base_dir = "BatchCases",
  parallel_xml = FALSE,
  workers = max(1, parallel::detectCores() - 1)
)
}
\description{
Resumable BatchCases: fetch → XML → JSON, all saved under BatchCases/\if{html}{\out{<search>}}/
}
