% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchdockets.R
\name{bd_fetch}
\alias{bd_fetch}
\title{Plain BatchDockets fetch (IDs only)}
\usage{
bd_fetch(
  q,
  max_n = Inf,
  page_sleep = 0,
  sleep_after = 0,
  endpoint = "/BatchDockets",
  retry = 3,
  retry_wait = 1,
  verbose = TRUE
)
}
\description{
Plain BatchDockets fetch (IDs only)
}
