% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchdockets.R
\name{bd_run_full_resumable}
\alias{bd_run_full_resumable}
\title{Resumable BatchDockets: IDs → XML → JSON
Saves to BatchDockets/\if{html}{\out{<search>}}/
JSON file will be \if{html}{\out{<search>}}_id_keyed.json}
\usage{
bd_run_full_resumable(
  q,
  page_sleep = 0,
  sleep_after = 0,
  endpoint = "/BatchDockets",
  retry = 3,
  retry_wait = 1,
  verbose = TRUE,
  parallel_xml = FALSE,
  workers = max(1, parallel::detectCores() - 1),
  base_dir = "BatchDockets"
)
}
\description{
Resumable BatchDockets: IDs → XML → JSON
Saves to BatchDockets/\if{html}{\out{<search>}}/
JSON file will be \if{html}{\out{<search>}}_id_keyed.json
}
