% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchnews.R
\name{bn_fetch}
\alias{bn_fetch}
\title{Plain in-memory BatchNews fetch (original)}
\usage{
bn_fetch(
  q,
  max_n = Inf,
  page_sleep = 0,
  sleep_after = 0,
  endpoint = "/BatchNews",
  retry = 2,
  retry_wait = 1,
  verbose = TRUE
)
}
\description{
Plain in-memory BatchNews fetch (original)
}
