% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchnews.R
\name{bn_fetch_resumable_auto}
\alias{bn_fetch_resumable_auto}
\title{Resumable / checkpointed BatchNews fetch
Saves to BatchNews/\if{html}{\out{<search>}}/...}
\usage{
bn_fetch_resumable_auto(
  q,
  max_n = Inf,
  page_sleep = 0,
  sleep_after = 0,
  endpoint = "/BatchNews",
  retry = 2,
  retry_wait = 1,
  verbose = TRUE,
  base_dir = "BatchNews"
)
}
\description{
Resumable / checkpointed BatchNews fetch
Saves to BatchNews/\if{html}{\out{<search>}}/...
}
