% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcases.R
\name{cases_to_idkeyed_json}
\alias{cases_to_idkeyed_json}
\title{Convert rows with xml to id-keyed JSON}
\usage{
cases_to_idkeyed_json(
  df,
  id_col = "ResultId",
  xml_col = "xml",
  pretty = TRUE,
  auto_unbox = TRUE
)
}
\description{
Convert rows with xml to id-keyed JSON
}
