% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchdockets.R
\name{dockets_to_idkeyed_json}
\alias{dockets_to_idkeyed_json}
\title{turn docket rows with xml into id-keyed JSON}
\usage{
dockets_to_idkeyed_json(
  df,
  id_col = "ResultId",
  xml_col = "xml",
  pretty = TRUE,
  auto_unbox = TRUE
)
}
\description{
turn docket rows with xml into id-keyed JSON
}
