% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lexis_auth}
\alias{lexis_auth}
\title{Authenticate to LexisNexis with a client id/secret}
\usage{
lexis_auth(key, secret)
}
\arguments{
\item{key}{Client ID}

\item{secret}{Client secret}
}
\description{
Stores the OAuth token in a global and in LN_TOKEN env var
so other functions can reuse it.
}
