% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lexis_auth_env}
\alias{lexis_auth_env}
\title{Authenticate using env vars LEXIS_CLIENT_ID / LEXIS_CLIENT_SECRET}
\usage{
lexis_auth_env()
}
\description{
Authenticate using env vars LEXIS_CLIENT_ID / LEXIS_CLIENT_SECRET
}
