% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ln_get_json}
\alias{ln_get_json}
\title{GET JSON with bearer and basic retry}
\usage{
ln_get_json(
  endpoint_or_full,
  query = NULL,
  retry = 3,
  retry_wait = 1,
  sleep_after = 0,
  ua = "lexisbatch/R"
)
}
\description{
GET JSON with bearer and basic retry
}
\keyword{internal}
